数字

+++ 0007.ReverseInteger/reverseInteger.cpp
@@ -3,7 +3,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        long res = 0;
+        long long res = 0;
         while (x != 0){
             res = res*10 + x%10;
             x /= 10;
